update with-loading example
