+++ examples/with-loading/pages/about.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react'
-import Header from '../components/Header'
 
 export default class About extends Component {
   // Add some delay
@@ -11,11 +10,6 @@ export default class About extends Component {
   }
 
   render () {
-    return (
-      <div>
-        <Header />
-        <p>This is about Next!</p>
-      </div>
-    )
+    return <p>This is about Next!</p>
   }
 }
