+++ examples/with-loading/pages/forever.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react'
-import Header from '../components/Header'
 
 export default class Forever extends Component {
   // Add some delay
@@ -11,11 +10,6 @@ export default class Forever extends Component {
   }
 
   render () {
-    return (
-      <div>
-        <Header />
-        <p>This page was rendered for a while!</p>
-      </div>
-    )
+    return <p>This page was rendered for a while!</p>
   }
 }
