+++ examples/with-loading/pages/index.js
@@ -1,9 +1,3 @@
 import React from 'react'
-import Header from '../components/Header'
 
-export default () => (
-  <div>
-    <Header />
-    <p>Hello Next!</p>
-  </div>
-)
+export default () => <p>Hello Next!</p>
